+++ test/localization.test.ts
@@ -1,4 +1,4 @@
-import { Enum, type IEnum } from '../src';
+import { DefaultLocalize, Enum, type IEnum } from '../src';
 import {
   genSillyLocalizer,
   localeCN,
@@ -14,45 +14,68 @@ describe('should support localization', () => {
   test('should show English by default', () => {
     const weekEnum = Enum(StandardWeekConfig);
     testEnum(weekEnum, localeEN);
+    testBuiltInResources(weekEnum, localeEN);
   });
 
   test('should show Chinese after changing lang', () => {
     setLang('zh-CN');
     const weekEnum = Enum(StandardWeekConfig);
     testEnum(weekEnum, localeCN);
+    testBuiltInResources(weekEnum, localeCN);
   });
 
   test('should show English after changing back', () => {
     setLang('en-US');
     const weekEnum = Enum(StandardWeekConfig);
     testEnum(weekEnum, localeEN);
+    testBuiltInResources(weekEnum, localeEN);
   });
 
   test('should show original label if no localization found', () => {
     setLang(undefined);
     const weekEnum = Enum(StandardWeekConfig);
     testEnum(weekEnum, noLocale);
+    testBuildInResourcesWithDefaultImp(weekEnum);
+  });
+
+  test('should show original label if Enum.localize is explicitly set to undefined ', () => {
+    setLang(undefined);
+    Enum.localize = undefined as any;
+    const weekEnum = Enum(StandardWeekConfig);
+    testEnum(weekEnum, noLocale);
+    testBuiltInResources(weekEnum, noLocale);
   });
 
   test('should respect Enum options over global setting (Chinese over English)', () => {
     setLang('en-US');
     const weekEnum = Enum(StandardWeekConfig, { localize: genSillyLocalizer('zh-CN') });
     testEnum(weekEnum, localeCN);
+    testBuiltInResources(weekEnum, localeCN);
   });
   test('should respect Enum options over global setting (undefined over English)', () => {
     setLang('en-US');
     const weekEnum = Enum(StandardWeekConfig, { localize: undefined });
     testEnum(weekEnum, localeEN);
+    testBuiltInResources(weekEnum, localeEN);
   });
   test('should respect Enum options over global setting (Chinese over undefined)', () => {
     setLang(undefined);
     const weekEnum = Enum(StandardWeekConfig, { localize: genSillyLocalizer('zh-CN') });
     testEnum(weekEnum, localeCN);
+    testBuiltInResources(weekEnum, localeCN);
   });
   test('should respect Enum options over global setting (both undefined)', () => {
     setLang(undefined);
     const weekEnum = Enum(StandardWeekConfig, { localize: undefined });
     testEnum(weekEnum, noLocale);
+    testBuildInResourcesWithDefaultImp(weekEnum);
+  });
+  test('should respect Enum options over global setting (both explicit undefined)', () => {
+    setLang(undefined);
+    Enum.localize = undefined as any;
+    const weekEnum = Enum(StandardWeekConfig, { localize: undefined });
+    testEnum(weekEnum, noLocale);
+    testBuiltInResources(weekEnum, noLocale);
   });
 });
 
@@ -92,4 +115,37 @@ function testEnum(
       return acc;
     }, {} as Record<number, { text: string | undefined }>)
   );
+}
+
+function testBuiltInResources(
+  weekEnum: IEnum<
+    typeof StandardWeekConfig,
+    keyof typeof StandardWeekConfig,
+    typeof StandardWeekConfig[keyof typeof StandardWeekConfig]['value']
+  >,
+  locales: typeof localeEN | typeof localeCN | typeof noLocale
+) {
+  const withDefaultFirstOption = weekEnum.options({ firstOption: true });
+  expect(withDefaultFirstOption).toHaveLength(8);
+  expect(withDefaultFirstOption[0]).toEqual({
+    value: '',
+    key: '',
+    label: locales['enum-plus.options.all'],
+  });
+}
+
+function testBuildInResourcesWithDefaultImp(
+  weekEnum: IEnum<
+    typeof StandardWeekConfig,
+    keyof typeof StandardWeekConfig,
+    typeof StandardWeekConfig[keyof typeof StandardWeekConfig]['value']
+  >
+) {
+  const withDefaultFirstOption = weekEnum.options({ firstOption: true });
+  expect(withDefaultFirstOption).toHaveLength(8);
+  expect(withDefaultFirstOption[0]).toEqual({
+    value: '',
+    key: '',
+    label: DefaultLocalize('enum-plus.options.all'),
+  });
 }